HtmlBridge P3.11: extract fetch/XMLHttpRequest into a networking module
Continue Phase 3 of the HtmlBridge complexity-reduction roadmap by turning the
fetch / XMLHttpRequest surface into a co-located feature binding module.

FetchBinding (Broiler.HtmlBridge.Dom.Features, split across FetchBinding.cs,
FetchBinding.Callbacks.cs and FetchBinding.Xhr.cs to stay under the
750-line-per-file guideline) now owns the whole fetch polyfill and its
Headers/Request/Response/FormData/Blob/AbortController helper objects, the
Response static factories (new Response / Response.json / Response.redirect) and
the XMLHttpRequest polyfill. This moves RegisterFetchAndHttpApis (now Install),
the four fetch Js...Core callbacks out of the shared 1516-line registration
callback file, the four fetch delegate types out of the shared Common.cs, and
RegisterXMLHttpRequest.

Host I/O goes through the injected Phase 2 ResourceLoader (the module holds a
reference to it), so no feature callback constructs an HttpClient -- the seam
Phase 7 builds on. The only remaining bridge coupling, the page URL used to
resolve a relative Response.redirect target, is the narrow IFetchHost.PageUrl,
implemented via one explicit interface member in DomBridge.FetchHost.cs so the
public surface is unchanged.

Two non-networking registrations that historically lived inside the fetch
registration -- MessageChannel (messaging) and getComputedStyle (CSSOM) -- were
relocated to the window-globals registration site. The caller now does
_fetch.Install(context, window); the old Registration/Fetch.cs and
Registration/XmlHttpRequest.cs are deleted.

Behaviour-preserving; no public-API change (module and contract are internal).
Tests: Broiler.Cli.Tests/FetchBindingModuleTests.cs (co-location / host-contract
/ ResourceLoader-ownership guards plus Response, Headers/FormData, XHR-installed
and relocated-MessageChannel/getComputedStyle characterizations, all
network-free). Regression check vs the P3.10 baseline: the network,
computed-style, messaging and selector suites pass unchanged (286 tests); the
pre-existing environmental failures (assembly-reflection migration checks,
iframe sub-document HTTP, real-HTTP fetch body readers, and the :root/:lang
selector tests) fail identically on both sides.
